move and add tests
